requirements.txt
Django==5.0.1
channels==4.0.0
django-notifications-hq==1.8.3
Pillow==10.2.0
python-dotenv==1.0.1